requirements.txt
google-generativeai
python-dotenv
mesa
networkx
pandas
matplotlib
seaborn
streamlit
openpyxl
plotly